trading_engine/requirements.txt
asgiref==3.7.2
Django==4.2.10
django-environ==0.11.2
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
aioredis==2.0.1
django-redis==5.4.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
alpaca-trade-api==3.0.2
redis==5.0.1
sortedcontainers==2.4.0
sqlparse==0.4.4
typing_extensions==4.10.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-ratelimit==4.1.0
pandas==2.1.4
numpy==1.26.3
